requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
sqlalchemy>=2.0
pydantic[email]>=2.5
python-dotenv>=1.0
email-validator>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
